+++ CHANGELOG.md
@@ -1,15 +1,29 @@
 # Change Log
 
-## 1.1.0 & 2.1.0 (2021-11-25)
+## 3.0.0 (2023 Mar 10)
+
 ### New
+
+Vue 3 support by default!
+
+### Fixed
+
+- Fixed navigating list items with keys when disabled item prop is set
+
+## 1.1.0 & 2.1.0 (2021-11-25)
+
+### New
+
 - [Custom Event Listeners](./README.md#custom-event-listeners) support
 
 ## 1.0.1 (2021-10-8)
+
 ### Fixed
 
-- Warning when compiling: `"export 'defineComponent' was not found in 'vue'` [Issue 24](https://github.com/sssmi/vue-combo-blocks/issues/24) 
+- Warning when compiling: `"export 'defineComponent' was not found in 'vue'` [Issue 24](https://github.com/sssmi/vue-combo-blocks/issues/24)
 
 ## 1.0.0 (2021-7-6)
+
 ### New
 
 Vue 3 support when using @next tag:
@@ -17,11 +31,14 @@ Vue 3 support when using @next tag:
 `npm i vue-combo-blocks@next`
 
 ### Changed
+
 **Removed:**
+
 - `ItemMouseLeave` state change type
 - Dropped support for IE11. (Unless you use babel, or similar to polyfill Object.assign)
 
 **Added:**
+
 - `MenuMouseLeave` state change type
 - `circular` prop controls what happens when navigation with arrow keys and list bottom or top is reached.
 
@@ -30,6 +47,7 @@ Vue 3 support when using @next tag:
 ### New
 
 - Use terser for minified build.
+
 ## 0.4.0 (2021-05-5)
 
 ### New